+++ gem5/src/mem/ruby/network/garnet2.0/GarnetNetwork.py
@@ -76,7 +76,7 @@ class GarnetNetwork(RubyNetwork):
     draino = Param.Bool(False, "enable draino simulation")
     draino_freq = Param.UInt32(5, "run draino every N drain cycles")
     draino_idle_cycles = Param.UInt32(10, "once in idle, draino will wait this many cycles before continuing.")
-    draino_latency_threshold = Param.UInt32(1, "draino will not react to a change in latency less than this magnitude.")
+    draino_latency_threshold = Param.Float(1, "draino will not react to a change in latency less than this magnitude.")
 
 class GarnetNetworkInterface(ClockedObject):
     type = 'GarnetNetworkInterface'
